Refactor createPages()
